Evaluate on subsets [Resolves #535, #138]
This commit adds support for evaluating models against subsets of their
predictions, in both training and testing. It adds three tables to the
results schemas to track subsets and their evaluations:

  - `model_metadata.subsets` stores subset metadata, including a hash,
    the subset configuration, and the time the row was created
  - `train_results.subset_evaluations` and
    `test_results.subset_evaluations` store evaluations for each subset

A new alembic upgrade script creates the subsets tables.

Testing factories are included for the subsets and subset_evaluations
tables, and a test for the factories ensures that the foreign keys in
the subset_evaluations tables are correctly configured.

Most of the remaining code changes are made to the ModelEvaluator class,
which can now process subset queries and write the results to the
appropriate table [#535] and will record `NULL` values for undefined
metrics (whether due to an empty subset or lack of variation in labels
[#138]).

However, some changes are made elsewhere in the experiment to allow
(optionally) including subsets in the experiment configuration file,
including storing subset metadata in the `model_metadata.subsets` table
and iterating over subsets in the model tester.

In addition, some changes to the documentation and `.gitignore` are
included to make modifying the results schema more joyful.